# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GM Startd API documentation",
        "title": "GM Startd API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/healthcheck": {
            "get": {
                "description": "Health checking for the service",
                "produces": [
                    "text/plain"
                ],
                "operationId": "HealthCheckHandler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Get Users",
                "produces": [
                    "application/json"
                ],
                "operationId": "UsersHandler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Users"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entities.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Response Code",
                    "type": "string"
                },
                "error": {
                    "description": "Response Error",
                    "type": "string"
                }
            }
        },
        "entities.User": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "UserId",
                    "type": "integer"
                },
                "name": {
                    "description": "UserName",
                    "type": "string"
                }
            }
        },
        "entities.Users": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Users",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.User"
                    }
                }
            }
        }
    }
}